M081: every item number the package ships is an integer (#88)
* M081 T1: read item numbers as integers in the four data-raw builders

* M081 T2, T3: sweep every shipped dataset for double item numbers, and pin the retype against the merge base

* M081 T4: drop the seven coercions the double item numbers required

* M081 T5: record the artifact and module round-trip evidence

* M081 T6: document the integer item numbers in the datasets, DESIGN, NEWS and DECISIONS

* review M081: fresh evidence for six criteria, and four fixes from the diff review

The review fixes:
- write_module() compared item_order against module$items with identical(),
  coercing only one side, so a module whose items are doubles -- one saved to
  .rds under 0.2.0 -- was refused where the merge base wrote it. The comparison
  is now by value and by length; two tests cover it.
- test-module_file.R and test-module.R wrapped as.integer() or expect_equal()
  around item comparisons, so nothing pinned hitop_module()$items to integer.
- The merge-base test's skip guard probed a column that is NULL on the four
  *_scales objects; it now checks whether the retype is a no-op.
- NEWS' arithmetic claim names the limit it holds under.

+++ R/rename_hitopsr_items.R
@@ -40,7 +40,7 @@ rename_hitopsr_items <- function(
   validate_string(prefix, arg = "prefix")
 
   ## Track matched standard item numbers for the final summary warning
-  matched_hsr <- numeric(0)
+  matched_hsr <- integer(0)
 
   if (method == "original") {
     data_cols <- colnames(data)
